flask-server/requirements.txt
flask
qrcode
pillow
flask-cors
opencv-python
pyzbar
pyzbar[scripts]
Werkzeug
cryptography
requests
